Make emojis that start with a dash (e.g. -1) work
docopt takes '-1' as an option even though there is no config for it,
so I ripped out docopt and replaced with argparse which does give us
a way to handle this by defining -s, -h/--help, and --no-copy and have
argparse give us all the unknown args.  When then have to whitelist them
in the cleanup process.